+++ src/hooks/useSelectLoanProduct.ts
@@ -1,0 +1,32 @@
+import {useMutation} from '@tanstack/react-query';
+import {useRouter} from 'next/navigation';
+
+import {loanProductApi} from '@/apis/loanProducts';
+
+export const useSelectLoanProduct = () => {
+    const router = useRouter();
+
+    return useMutation({
+        mutationFn: async (productId: number) => {
+            const token = localStorage.getItem('accessToken');
+            if (!token) {
+                throw {response: {status: 401}};
+            }
+            return loanProductApi.selectLoanProduct(productId, token);
+        },
+        onSuccess: () => {
+            router.push('/credit-evaluation');
+        },
+        onError: (error: any) => {
+            console.error('대출 신청 에러:', error);
+
+            if (error?.response?.status === 401) {
+                alert('인증에 실패했습니다. 다시 로그인해주세요.');
+                localStorage.removeItem('accessToken');
+                router.push('/auth/login');
+            } else {
+                alert(`대출 신청에 실패했습니다: ${error?.response?.data?.message || '알 수 없는 오류'}`);
+            }
+        },
+    });
+};
